camera.d: fixed so that if faces origo from the south game.d: Fixed control / space actions main.d: made so that the garbage collector cant run on exit. fuck oyu damn dirrlicht-exceptions! pos.d: Made fix on graphregionnum.max (When used with ranges, should be exclusive on upper limit) renderer.d: awsome, denrers stuffs!! VBOMaker.d: works!! world.d: stuff check changes worldgen.d: blarp etc

+++ gameroot/shaders/renderGR.vert
@@ -5,7 +5,7 @@ const float pixelWidth = 1.0/1024.0;
 const vec2 tileSize = vec2(16.0, 16.0) * pixelWidth;
 
 uniform mat4 VP;
-uniform ivec3 offset; 
+uniform vec3 offset; 
 
 in vec3 position;
 in int type;
@@ -13,9 +13,6 @@ in int type;
 out vec2 tex_texcoord;
    
 void main(){
-
-   //tex_texcoord = tileSize*tileTexIdx + in_texcoord;
-
    gl_Position = VP * vec4(position+offset, 1.0);
 }
